refactor(dapp): slice provider logic (#149)
* wip: mid staking logic extraction

* wip: use-staking

* wip: update use-staking

* wip: add missing methods to use-faith

* wip: completed use-staking

* wip: added props to use-faith and use-staking

* wip: added use-swap

* wip: delete unused Exit and Enter pages

* wip: updateWallet refactor

* wip: build successful

* wip: refreshWalletState refactor

* wip: added AppProvider as provider aggregator

* wip: fix to use-refreshable-account-metrics

* wip: removed unnecessary ts-ignore for ethereum obj
