+++ apps/dapp/src/components/AMM/Buy.tsx
@@ -1,6 +1,6 @@
-import { Input } from 'components/Input/Input';
+import React, { FC, useEffect, useState, useCallback } from 'react';
 import { BigNumber } from 'ethers';
-import React, { FC, useEffect, useState } from 'react';
+import { Input } from 'components/Input/Input';
 import { formatNumber } from 'utils/formatter';
 import {
   ConvoFlowTitle,
@@ -14,6 +14,8 @@ import Slippage from 'components/Slippage/Slippage';
 import { Button } from 'components/Button/Button';
 import { TICKER_SYMBOL } from 'enums/ticker-symbol';
 import { useWallet } from 'providers/WalletProvider';
+import { useSwap } from 'providers/SwapProvider';
+import { useRefreshWalletState } from 'hooks/use-refresh-wallet-state';
 import { fromAtto, toAtto } from 'utils/bigNumber';
 import { noop } from 'utils/helpers';
 
@@ -25,8 +27,8 @@ interface BuyProps {
 const ENV_VARS = import.meta.env;
 
 export const Buy: FC<BuyProps> = ({ onSwapArrowClick, small }) => {
-  const { balance, getBalance, updateWallet, buy, getBuyQuote, templePrice } =
-    useWallet();
+  const { balance, getBalance } = useWallet();
+  const { buy, getBuyQuote, templePrice } = useSwap();
 
   const [stableCoinAmount, setStableCoinAmount] = useState<number | ''>('');
   const [stableCoinWalletAmount, setStableCoinWalletAmount] =
@@ -35,6 +37,8 @@ export const Buy: FC<BuyProps> = ({ onSwapArrowClick, small }) => {
   const [templeWalletAmount, setTempleWalletAmount] = useState<number>(0);
   const [slippage, setSlippage] = useState<number>(1);
   const [minAmountOut, setMinAmountOut] = useState<number>(0);
+
+  const refreshWalletState = useRefreshWalletState();
 
   const handleUpdateStableCoinAmount = async (value: number | '') => {
     setStableCoinAmount(value === 0 ? '' : value);
@@ -77,7 +81,7 @@ export const Buy: FC<BuyProps> = ({ onSwapArrowClick, small }) => {
 
   useEffect(() => {
     async function onMount() {
-      await updateWallet();
+      await refreshWalletState();
       setRewards('');
       setMinAmountOut(0);
     }
