+++ apps/dapp/src/components/AMM/Unlock.tsx
@@ -1,9 +1,10 @@
+import React, { FC, useEffect } from 'react';
 import ClaimOGTemple from 'components/AMM/helpers/ClaimOGTemple';
 import { Button } from 'components/Button/Button';
 import { Flex } from 'components/Layout/Flex';
 import Tooltip, { TooltipIcon } from 'components/Tooltip/Tooltip';
-import { useWallet } from 'providers/WalletProvider';
-import React, { FC, useEffect } from 'react';
+import { useStaking } from 'providers/StakingProvider';
+import { useRefreshWalletState } from 'hooks/use-refresh-wallet-state';
 import { TICKER_SYMBOL } from 'enums/ticker-symbol';
 import {
   ConvoFlowTitle,
@@ -18,16 +19,18 @@ interface UnlockProps {
 }
 
 export const Unlock: FC<UnlockProps> = ({ onExitClick, onReturnClick }) => {
-  const { updateWallet, lockedEntries, claimOgTemple } = useWallet();
+  const { lockedEntries, claimOgTemple } = useStaking();
+
+  const refreshWalletState = useRefreshWalletState();
 
   const handleClaimOgTemple = async (index: number) => {
     await claimOgTemple(index);
-    updateWallet();
+    refreshWalletState();
   };
 
   useEffect(() => {
     async function onMount() {
-      await updateWallet();
+      await refreshWalletState();
     }
 
     onMount();
