+++ apps/dapp/src/main.tsx
@@ -1,13 +1,13 @@
-import React from 'react';
+import React, { Suspense } from 'react';
 import ReactDOM from 'react-dom';
 import { BrowserRouter, Routes, Route, Navigate } from 'react-router-dom';
-import styled, { ThemeProvider } from 'styled-components';
+import styled from 'styled-components';
+import { GlobalStyle } from 'styles/GlobalStyle';
+
+import { AppProvider } from 'providers/AppProvider';
 
 import NotificationManager from 'components/Notification/NotificationManager';
-import { NotificationProvider } from 'providers/NotificationProvider';
-import { WalletProvider } from 'providers/WalletProvider';
-import { GlobalStyle } from 'styles/GlobalStyle';
-import { theme } from 'styles/theme';
+
 import PageLayout from 'components/Layouts/Page';
 import Loader from 'components/Loader/Loader';
 
@@ -17,12 +17,12 @@ import Disclaimer from 'components/Pages/Disclaimer';
 // Separate Chunks
 const AmmSpaRoot = React.lazy(() => import('components/Pages/AMM'));
 const DAppRoot = React.lazy(() => import('components/Pages/DAppRoot'));
-const FireRitualistCashback = React.lazy(() => import('components/Pages/FireRitualistCashback'));
+const FireRitualistCashback = React.lazy(
+  () => import('components/Pages/FireRitualistCashback')
+);
 const TeamPayments = React.lazy(() => import('components/Pages/TeamPayments'));
 const FaithAirdrop = React.lazy(() => import('components/Pages/FaithAirdrop'));
 const Claim = React.lazy(() => import('components/Pages/Claim'));
-const Exit = React.lazy(() => import('components/Pages/Exit'));
-const Enter = React.lazy(() => import('components/Pages/Enter'));
 
 const LoaderWrapper = styled.div`
   display: flex;
@@ -38,47 +38,55 @@ interface LazyPageProps {
 }
 
 const LazyPage = ({ component: Component }: LazyPageProps) => (
-  <React.Suspense fallback={(
-    <LoaderWrapper>
-      <Loader />
-    </LoaderWrapper>
-  )}>
+  <Suspense
+    fallback={
+      <LoaderWrapper>
+        <Loader />
+      </LoaderWrapper>
+    }
+  >
     <Component />
-  </React.Suspense>
+  </Suspense>
 );
 
 ReactDOM.render(
   <React.StrictMode>
-    <GlobalStyle />
-    <NotificationProvider>
-      <WalletProvider>
-        <ThemeProvider theme={theme}>
-          <BrowserRouter>
-            <Routes>
-              <>
-                <Route path="/the-temple" element={<LazyPage component={AmmSpaRoot} />} />
-                <Route path="/dapp" element={<LazyPage component={DAppRoot} />} />
-                <Route path="/" element={<PageLayout />}>
-                  <Route path="/" element={<Home />} />
-                  <Route path="disclaimer" element={<Disclaimer />} />
-                  <Route path="enter" element={<LazyPage component={Enter} />} />
-                  <Route path="exit" element={<LazyPage component={Exit} />} />
-                  <Route path="faith-airdrop" element={<LazyPage component={FaithAirdrop} />} />
-                  <Route
-                    path="fire-ritualist-apy-topup"
-                    element={<LazyPage component={FireRitualistCashback} />}
-                  />
-                  <Route path="temple-claim" element={<LazyPage component={Claim} />} />
-                  <Route path="team-payments" element={<LazyPage component={TeamPayments} />} />
-                  <Route path="/*" element={<Navigate replace to="/" />} />
-                </Route>
-              </>
-            </Routes>
-          </BrowserRouter>
-          <NotificationManager />
-        </ThemeProvider>
-      </WalletProvider>
-    </NotificationProvider>
+    <AppProvider>
+      <GlobalStyle />
+      <BrowserRouter>
+        <Routes>
+          <>
+            <Route
+              path="/the-temple"
+              element={<LazyPage component={AmmSpaRoot} />}
+            />
+            <Route path="/dapp" element={<LazyPage component={DAppRoot} />} />
+            <Route path="/" element={<PageLayout />}>
+              <Route path="/" element={<Home />} />
+              <Route path="disclaimer" element={<Disclaimer />} />
+              <Route
+                path="faith-airdrop"
+                element={<LazyPage component={FaithAirdrop} />}
+              />
+              <Route
+                path="fire-ritualist-apy-topup"
+                element={<LazyPage component={FireRitualistCashback} />}
+              />
+              <Route
+                path="temple-claim"
+                element={<LazyPage component={Claim} />}
+              />
+              <Route
+                path="team-payments"
+                element={<LazyPage component={TeamPayments} />}
+              />
+              <Route path="/*" element={<Navigate replace to="/" />} />
+            </Route>
+          </>
+        </Routes>
+      </BrowserRouter>
+      <NotificationManager />
+    </AppProvider>
   </React.StrictMode>,
   document.getElementById('root')
 );
