+++ apps/dapp/src/providers/env.ts
@@ -10,8 +10,6 @@ export const TEMPLE_STAKING_ADDRESS =
   ENV_VARS.VITE_PUBLIC_TEMPLE_STAKING_ADDRESS;
 export const TREASURY_ADDRESS = ENV_VARS.VITE_PUBLIC_TREASURY_ADDRESS;
 export const EXIT_QUEUE_ADDRESS = ENV_VARS.VITE_PUBLIC_EXIT_QUEUE_ADDRESS;
-export const OPENING_CEREMONY_ADDRESS =
-  ENV_VARS.VITE_PUBLIC_OPENING_CEREMONY_ADDRESS;
 export const TEMPLE_CASHBACK_ADDRESS =
   ENV_VARS.VITE_PUBLIC_TEMPLE_CASHBACK_ADDRESS;
 export const AMM_WHITELIST_ADDRESS =
@@ -68,7 +66,6 @@ if (
   TREASURY_ADDRESS === undefined ||
   LOCKED_OG_TEMPLE_ADDRESS === undefined ||
   EXIT_QUEUE_ADDRESS === undefined ||
-  OPENING_CEREMONY_ADDRESS === undefined ||
   TEMPLE_CASHBACK_ADDRESS === undefined ||
   TEMPLE_V2_ROUTER_ADDRESS === undefined ||
   TEMPLE_V2_PAIR_ADDRESS === undefined ||
@@ -85,7 +82,6 @@ TEMPLE_STAKING_ADDRESS=${TEMPLE_STAKING_ADDRESS}
 TREASURY_ADDRESS=${TREASURY_ADDRESS}
 LOCKED_OG_TEMPLE_ADDRESS=${LOCKED_OG_TEMPLE_ADDRESS}
 EXIT_QUEUE_ADDRESS=${EXIT_QUEUE_ADDRESS}
-OPENING_CEREMONY_ADDRESS=${OPENING_CEREMONY_ADDRESS}
 TEMPLE_CASHBACK_ADDRESS=${TEMPLE_CASHBACK_ADDRESS}
 TEMPLE_V2_ROUTER_ADDRESS=${TEMPLE_V2_ROUTER_ADDRESS}
 TEMPLE_V2_PAIR_ADDRESS=${TEMPLE_V2_PAIR_ADDRESS}
