+++ apps/dapp/src/providers/errors.ts
